# frontend/src/docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "SynqX ETL Agent",
    "description": "Universal ETL Engine",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:8000",
      "description": "Local server"
    },
    {
      "url": "https://api.synqx.io",
      "description": "Production server"
    }
  ],
  "paths": {
...
    "/api/v1/explorer/{connection_id}/execute": {
      "post": {
        "tags": ["data exploration"],
        "summary": "Execute Connection Query",
        "description": "Execute a raw query against a connection.\nSupports SQL and NoSQL depending on the connector type.",
        "operationId": "execute_connection_query_api_v1_explorer__connection_id__execute_post",
        "security": [
          {
            "OAuth2PasswordBearer": []
          },
          {
            "APIKeyHeader": []
          }
        ],
        "parameters": [
          {
            "name": "connection_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Connection Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/QueryRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/QueryResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/explorer/history": {
      "get": {
        "tags": ["data exploration"],
        "summary": "Get Query History",
        "description": "Get execution history for the active workspace.",
        "operationId": "get_query_history_api_v1_explorer_history_get",
        "security": [
          {
            "OAuth2PasswordBearer": []
          },
          {
            "APIKeyHeader": []
          }
        ],
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 50,
              "title": "Limit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 0,
              "title": "Offset"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/HistoryItem"
                  },
                  "title": "Response Get Query History Api V1 Explorer History Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": ["data exploration"],
        "summary": "Clear Query History",
        "description": "Clear execution history for the active workspace.",
        "operationId": "clear_query_history_api_v1_explorer_history_delete",
        "security": [
          {
            "OAuth2PasswordBearer": []
          },
          {
            "APIKeyHeader": []
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        }
      }
    },
    "/api/v1/explorer/{connection_id}/schema-metadata": {
      "get": {
        "tags": ["data exploration"],
        "summary": "Get Connection Schema Metadata",
        "description": "Get full schema metadata for autocompletion.",
        "operationId": "get_connection_schema_metadata_api_v1_explorer__connection_id__schema_metadata_get",
        "security": [
          {
            "OAuth2PasswordBearer": []
          },
          {
            "APIKeyHeader": []
          }
        ],
        "parameters": [
          {
            "name": "connection_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Connection Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/jobs/{job_id}": {
      "get": {
        "tags": ["jobs & runs"],
        "summary": "Get Job",
        "description": "Get detailed information about a specific job",
        "operationId": "get_job_api_v1_jobs__job_id__get",
        "security": [
          {
            "OAuth2PasswordBearer": []
          },
          {
            "APIKeyHeader": []
          }
        ],
        "parameters": [
          {
            "name": "job_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Job Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "HistoryItem": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "query": {
            "type": "string",
            "title": "Query"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "execution_time_ms": {
            "type": "integer",
            "title": "Execution Time Ms"
          },
          "row_count": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Row Count"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "connection_name": {
            "type": "string",
            "title": "Connection Name"
          },
          "created_by": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Created By"
          }
        },
        "type": "object",
        "required": [
          "id",
          "query",
          "status",
          "execution_time_ms",
          "row_count",
          "created_at",
          "connection_name",
          "created_by"
        ],
        "title": "HistoryItem"
      },
...
      "JobRead": {
        "properties": {
          "pipeline_id": {
            "type": "integer",
            "title": "Pipeline Id"
          },
          "pipeline_version_id": {
            "type": "integer",
            "title": "Pipeline Version Id"
          },
          "status": {
            "$ref": "#/components/schemas/JobStatus"
          },
          "retry_count": {
            "type": "integer",
            "title": "Retry Count",
            "default": 0
          },
          "max_retries": {
            "type": "integer",
            "title": "Max Retries",
            "default": 3
          },
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "celery_task_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Celery Task Id"
          },
          "correlation_id": {
            "type": "string",
            "title": "Correlation Id"
          },
          "retry_strategy": {
            "$ref": "#/components/schemas/RetryStrategy"
          },
          "retry_delay_seconds": {
            "type": "integer",
            "title": "Retry Delay Seconds"
          },
          "infra_error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Infra Error"
          },
          "worker_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Worker Id"
          },
          "queue_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Queue Name"
          },
          "execution_time_ms": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Execution Time Ms"
          },
          "started_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Started At"
          },
          "completed_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Completed At"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "pipeline_id",
          "pipeline_version_id",
          "status",
          "id",
          "celery_task_id",
          "correlation_id",
          "retry_strategy",
          "retry_delay_seconds",
          "infra_error",
          "worker_id",
          "queue_name",
          "execution_time_ms",
          "started_at",
          "completed_at",
          "created_at",
          "updated_at"
        ],
        "title": "JobRead"
      },
...
      "JobStatus": {
        "type": "string",
        "enum": ["pending", "queued", "running", "success", "failed", "retrying", "cancelled"],
        "title": "JobStatus"
      },
...
      "QueryRequest": {
        "properties": {
          "query": {
            "type": "string",
            "title": "Query"
          },
          "limit": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Limit",
            "default": 100
          },
          "offset": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Offset",
            "default": 0
          },
          "params": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Params"
          }
        },
        "type": "object",
        "required": ["query"],
        "title": "QueryRequest"
      },
      "QueryResponse": {
        "properties": {
          "results": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "title": "Results"
          },
          "count": {
            "type": "integer",
            "title": "Count"
          },
          "columns": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Columns"
          }
        },
        "type": "object",
        "required": ["results", "count", "columns"],
        "title": "QueryResponse"
      },
...
      "RetryStrategy": {
        "type": "string",
        "enum": ["none", "fixed", "exponential_backoff", "linear_backoff"],
        "title": "RetryStrategy"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": ["loc", "msg", "type"],
        "title": "ValidationError"
      },
...
    },
    "securitySchemes": {
      "OAuth2PasswordBearer": {
        "type": "oauth2",
        "flows": {
          "password": {
            "scopes": {},
            "tokenUrl": "/api/v1/auth/login"
          }
        }
      },
      "APIKeyHeader": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-Key"
      }
    }
  }
}